Create vcs.xml
Answer the questions
Given bracket sequence: [((())()(())]]
Can this sequence be considered correct?
Answer: No
Why?
We have an inequality in the number of brackets.
Either we must replace the first parenthesis with a square bracket, or we must replace the second from the end square bracket with a round bracket
Examples:
[[(())()(())]]

[((())()(()))]